fix(user): rename directory name for conflict
